Add results from run of CPAN River 3000 on Jun 14 2018.
In FreeBSD-11.1 VM 'perl-reporter-01', using perl-5.28.0-RC2.

This required several passes due to timeouts, which caused the
application to terminate, after which I had to re-calculate the CPAN
River 3000 to exclude the distros timing out.

As a consequence, the first pass to actually complete did not record any
JSON for distros -- the overwhelming majority -- which PASSed on any of
the 3 or 4 earlier passes.  So the only distros for which JSON is being
stored are those which got a non-PASS grade or which got PASS only on
the final pass.