Version of trigger-tad.pl used in perl-reporter-03 on Jul 18 2018.
In this version:

(i) 'use lib' had hard-coded path to a variant of
Test::Against::Dev.  This has been merged into master but not yet
released to CPAN.

TODO: Once I've done a new CPAN release and installed that
in perl-reporter-03, I'll need to remove the 'use lib' line from the
script.

(ii) Contains hard-coded path for the URI at which to fetch cpanm.  This
variant cpanm does not contain an 'exit' call using the return value of
doit() as the exit status -- which I think was causing premature exit of
trigger-tad.pl.

TODO: Add a command-line option for the cpanm URI.