Vagrantfile for OpenBSD-6.6 VM (perl-reporter-09)
The VM originally targeted by 'perl-reporter-09' turned out to be
OpenBSD-6.5; I wanted 6.6.  So I called 'vagrant destroy' on the 6.5 box
and did a 'vagrant init' to get a 6.6 box.  I then pointed the
'perl-reporter-09' symlink to the new box.

In 6.6, we no longer have to call for a specific version of 'screen'.
It now DTRT.