Use 'pkgin -y' to upgrade packages
This is analogous to 'pkg install -y' in the FreeBSD Vagrantfile SHELL
script.

This generally worked well, but there were some dependency ordering
problems.  Most notably, 'vim-share' could not be upgraded and as a
result 'vim' was not only not upgraded but was apparently removed.
Internet search indicated that the immediate fix was (as vagrant user
inside the VM): 'sudo pkg_admin rebuild-tree'.  This worked, except
there may still be problems with the ruby-subversion package.